+++ test/package.test.js
@@ -14,7 +14,7 @@ describe('@authnomicon/oauth2', function() {
       
       expect(json.assembly.components).to.have.length(15);
       expect(json.assembly.components).to.include('authorization/service');
-      expect(json.assembly.components).to.include('token/http/service');
+      expect(json.assembly.components).to.include('token/service');
     });
   });